@@ -2,18 +2,18 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 
     <ToolCommandName>Web.Ping</ToolCommandName>
 
     <PackageId>Web.Ping</PackageId>
-    <PackageVersion>0.0.3</PackageVersion>
+    <PackageVersion>0.0.4</PackageVersion>
     <Authors>ZKEASOFT</Authors>
     <Description>Determine whether a remote web server is accessible over the network.</Description>
 
     <PackAsTool>True</PackAsTool>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Version>0.0.3</Version>
+    <Version>0.0.4</Version>
     <Copyright>ZKEASOFT</Copyright>
     <PackageProjectUrl>https://www.zkea.net</PackageProjectUrl>
     <RepositoryUrl>https://github.com/SeriaWei/Ping</RepositoryUrl>
@@ -24,10 +24,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="3.0.0" />
+    <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Include="Images\icon.png" Pack="true" PackagePath=""/>
+    <None Include="Images\icon.png" Pack="true" PackagePath="" />
   </ItemGroup>
 </Project>
